requirements.txt
mysql-connector-python==8.2.0
bcrypt==4.1.1

